+++ core-bundle/src/Command/UserPasswordCommand.php
@@ -11,7 +11,6 @@
 namespace Contao\CoreBundle\Command;
 
 use Contao\Config;
-use Contao\Encryption;
 use Patchwork\Utf8;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Exception\InvalidArgumentException;
@@ -148,9 +147,6 @@ class UserPasswordCommand extends ContainerAwareCommand
             );
         }
 
-        /** @var Encryption $encryption */
-        $encryption = $framework->getAdapter(Encryption::class);
-
-        return $encryption->hash($password);
+        return password_hash($password, PASSWORD_DEFAULT);
     }
 }